Updated Unity Version Fixed a cyclical method call whe healing on increasing health with powerups Fixed moon ordering

+++ Assets/_Scripts/Health/Health.cs
@@ -58,7 +58,7 @@ public class Health : MonoBehaviour, IDamageable
 
     public int Heal(int heal)
     {
-        Heal(heal);
+        Heal((float)heal);
         return (int)_health;
     }
 }
